[AMQ-8354] Extract ReplicaBaseObjectStorage. Small method refactoring. Add the missing deinitialization of Resynchronization storage.
